split post component
